Added a UberLazyHeapNode that extends HeapNode to keep track of the refChain variable of the original object instantiated. This allows us to intitialize any new symbolic fields in the class with the appropriate tags.
Added the initialization of the instance and static fields when a new concretization of the object was done in the UberLazyHelper, without this, the fields of the new concrete object were not symbolic.

InvokeVirtual bytecode for the uberlazy should be complete

Added a redundant execute method in the bytecodeUtils to get around the two choice generator issue. Need to fix this problem.