Uberlazy updates:
In the getfield class, the attribute as the name of the field is added on the stack. When the field is accessed the first time; there are still two choices: null and new. The second is a single partitionchoice which does not really create any branching but just allows us to add to the stack the relevant value in the attribute. This change caused us to check in the ifnull and ifnonnull bytecodes, whether the attribute is relevant to the uberlazy algorithm, aka not an instance of an \"Expression\".  When such an instance is discovered the concrete bytecode is executed.

The instruction factory insnClass in both uberlazy and symbc is initialised to the 260 values. Easier to have a constant value rather than keep changing as we add the bytecodes.